# FuzzDiff Performance improvements