--- a/src/tests/class_tests/openms/source/StopWatch_test.cpp
+++ b/src/tests/class_tests/openms/source/StopWatch_test.cpp
@@ -39,15 +39,19 @@
 
 #include <OpenMS/SYSTEM/StopWatch.h>
 
-#include <ctime>
+#include <chrono>
 /////////////////////////////////////////////////////////////
 
 using namespace OpenMS;
 
-void wait (int seconds)
+void wait(double seconds)
 {
-    clock_t endwait = clock () + seconds * CLOCKS_PER_SEC;
-    while (clock () < endwait) {}
+  auto start = std::chrono::system_clock::now();
+  while (true)
+  {
+   double s = std::chrono::duration<double>(std::chrono::system_clock::now() - start).count();
+   if (s > seconds) break;
+  };
 }
 
 START_TEST(StopWatch, "$Id$")
@@ -55,31 +59,41 @@ START_TEST(StopWatch, "$Id$")
 /////////////////////////////////////////////////////////////
 
 START_SECTION((StopWatch& operator = (const StopWatch& stop_watch)))
-  StopWatch s1;
-  s1.start();
-  wait(1);
-  s1.stop();
-
-  StopWatch s2;
-  TEST_EQUAL(s1!=s2, true)
-  s2 = s1;
-  TEST_EQUAL(s1==s2, true)
-
+  NOT_TESTABLE; // tested below
 END_SECTION
 
 START_SECTION((StopWatch()))
-  NOT_TESTABLE; // tested above
+  NOT_TESTABLE; // tested below
 END_SECTION
 
 START_SECTION((StopWatch(const StopWatch& stop_watch)))
-  StopWatch s1;
+  StopWatch s1, s2;
   s1.start();
-  wait(1);
+  wait(0.01);
+  TEST_EQUAL(s1 != s2, true) // before stop
+  s1.stop();
+  TEST_EQUAL(s1 != s2, true)
+  s2 = s1;
+  TEST_EQUAL(s1 == s2, true)
+
+  StopWatch s3(s1);
+  TEST_EQUAL(s1==s3, true)
+  
+  StopWatch s4;
+  s1.reset();
+  TEST_EQUAL(s1 == s4, true)
+
+  s1.start();
+  s2.start();
+  wait(0.01);
   s1.stop();
 
-  StopWatch s2(s1);
-  TEST_EQUAL(s1==s2, true)
-  
+  wait(0.01);
+  s2.stop();
+
+  TEST_EQUAL(s1 != s2, true)
+  TEST_EQUAL(s1 <= s2, true)
+  TEST_EQUAL(s2 >= s1, true)
 END_SECTION
 
 START_SECTION((bool isRunning() const))
@@ -87,40 +101,11 @@ START_SECTION((bool isRunning() const))
 
   w.start();
   TEST_EQUAL(w.isRunning(), true);
-
   w.stop();
-
 END_SECTION
 
 START_SECTION((bool operator != (const StopWatch& stop_watch) const))
-  StopWatch s, s2;
-  StopWatch s3(s2);
-  TEST_EQUAL(s2==s3, true);
-
-
-  s.start();
-  s2.start();
-  wait(3);
-  s.stop();
-
-  wait(3);
-  s2.stop();
-
-  TEST_EQUAL(s!=s2, true)
-
-  TEST_EQUAL(s<=s2, true)
-
-  TEST_EQUAL(s2>=s, true)
-
-  TEST_EQUAL(s2!=s3, true)
-  s3 = s2;
-  TEST_EQUAL(s2==s3, true)
-
-  s2.start();
-  wait(1);
-  TEST_EQUAL(s2==s3, false)
-  s2.stop();
-
+  NOT_TESTABLE; // tested above
 END_SECTION
 
 START_SECTION((bool operator < (const StopWatch& stop_watch) const))
@@ -148,41 +133,39 @@ START_SECTION((bool start()))
 END_SECTION
 
 START_SECTION((bool stop()))
+  const double t_wait = 0.2;
   StopWatch s;
   s.start();
-  wait(3);
+  wait(t_wait);
   s.stop();
 
-  TEST_EQUAL(s.getClockTime() > 2, true)
-  TEST_EQUAL(s.getClockTime() < 4, true)
+  TEST_EQUAL(s.getClockTime() > 0.1, true)
+  TEST_EQUAL(s.getClockTime() < 0.3, true)
   
   double t1 = s.getCPUTime();
   double t2 = s.getClockTime();
   double t3 = s.getSystemTime();
   double t4 = s.getUserTime();
   // wait some more
-  wait(3);
+  wait(0.1);
   // ... and see if time is still the old one
   TEST_EQUAL(s.getCPUTime(), t1)
   TEST_EQUAL(s.getClockTime(), t2)
   TEST_EQUAL(s.getSystemTime(), t3)
   TEST_EQUAL(s.getUserTime(), t4)
+  TEST_EQUAL(s.getCPUTime(), t1)
 
+  TEST_EQUAL(s.getCPUTime() > t_wait / 2, true) // waiting costs CPU time in our implementation... just not sure how much...
+  TEST_EQUAL(s.getClockTime() > t_wait * 0.95, true) // and must consume wall time
+  TEST_EQUAL(s.getClockTime() < t_wait * 3, true) // be a bit more loose if e.g. a VM is busy
+  TEST_EQUAL(s.getUserTime() > t_wait / 2, true) //  and some user time
+  TEST_EQUAL(s.getUserTime() < t_wait * 2, true)
+  TEST_EQUAL(s.getSystemTime() < t_wait, true) // and usually quite few system time
+                                               //(not guaranteed on VMs, therefore do a trivial check)
 END_SECTION
 
 START_SECTION((double getCPUTime() const ))
-  StopWatch s;
-  s.start();
-  wait(3);
-  s.stop();
-
-  TEST_EQUAL(s.getCPUTime() > 0.1, true) // waiting costs CPU time... just not sure how much...
-  TEST_EQUAL(s.getClockTime() > 2, true) // and must consume wall time
-  TEST_EQUAL(s.getClockTime() < 10, true) // be a bit more loose if e.g. a VM is busy
-  TEST_EQUAL(s.getUserTime() > 0.1, true) //  and some user time
-  TEST_EQUAL(s.getUserTime() < 4, true)
-  TEST_EQUAL(s.getSystemTime() < 3, true) // and usually quite few system time
-  //(not guaranteed on VMs, therefore do a trivial check)
+  NOT_TESTABLE; // done above
 END_SECTION
 
 START_SECTION((double getClockTime() const ))