## FuzzyDiff speed
comparing "large" files (6MB) using FuzzyDiff takes about 9 seconds on VS2015 Release(!), whereas its rather fast (0.3 sec) on GCC.
This PR speeds up FuzzyDiff by avoiding the  `stringstream>>double operator` which is awfully slow on VS.
Test time on Windows is now also 0.3 secs for this special case.
In general the combined time of TOPP+UTIL tests (1117 tests), many of which use FuzzyDiff, went from 55 sec to 36 sec (10 jobs parallel) on VS. GCC builds also profit, but not as much (102 sec --> 82 sec).

## reduces time for the stopwatch test 
12 sec to 0.5 sec by reducing waiting time