@timosachsenberg 2019-01-10
@jpfeuffer  Do you think that decreasing the wait time will cause errors?