@jpfeuffer 2019-01-10
Well the StopWatch test is very fragile and can only be tested roughly. I think he adapted the expected results accordingly and therefore I think it should be fine.